마우스 입력 여부 확인방식 비트플래그로 변경

+++ AR41Engine/Include/UI/UIWindow.cpp
@@ -9,7 +9,7 @@ CUIWindow::CUIWindow()  :
 {
 	m_WindowTypeName = "UIWindow";
 	m_Size = Vector2((float)CDevice::GetInst()->GetResolution().Width,
-		CDevice::GetInst()->GetResolution().Height);
+		(float)CDevice::GetInst()->GetResolution().Height);
 }
 
 CUIWindow::CUIWindow(const CUIWindow& Window)	:
